@@ -3,7 +3,6 @@ package p32929.easypasscodelock.Activities;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
@@ -11,17 +10,19 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.core.app.ActivityCompat;
+
 import p32929.easypasscodelock.Interfaces.ActivityChanger;
 import p32929.easypasscodelock.R;
 import p32929.easypasscodelock.Utils.EasyLock;
-import p32929.easypasscodelock.Utils.FayazSP;
+import p32929.easypasscodelock.Utils.EasylockSP;
 import p32929.easypasscodelock.Utils.LockscreenHandler;
 
 public class LockscreenActivity extends LockscreenHandler implements ActivityChanger {
 
+    @SuppressWarnings("rawtypes")
     private static Class classToGoAfter;
-    String tempPass = "";
-    private int[] passButtonIds = {
+    private final int[] passButtonIds = {
             R.id.lbtn1,
             R.id.lbtn2,
             R.id.lbtn3,
@@ -33,93 +34,87 @@ public class LockscreenActivity extends LockscreenHandler implements ActivityCha
             R.id.lbtn9,
             R.id.lbtn0
     };
-    private TextView textViewDot, textViewHAHA, textViewForgotPassword;
-    private Button buttonTick;
-    private ImageButton imageButtonDelete;
-    private RelativeLayout relativeLayoutBackground;
-
+    //
+    private final String checkStatus = "check";
+    private final String setStatus = "set";
+    private final String setStatus1 = "set1";
+    private final String changeStatus = "change";
+    private final String changeStatus1 = "change1";
+    private final String changeStatus2 = "change2";
+    String tempPass = "";
+    private TextView textViewDot;
+    private TextView textViewHAHA;
     private String passString = "", realPass = "";
     private String status = "";
-    //
-    private String checkStatus = "check";
-    private String setStatus = "set";
-    private String setStatus1 = "set1";
-    private String disableStatus = "disable";
-    private String changeStatus = "change";
-    private String changeStatus1 = "change1";
-    private String changeStatus2 = "change2";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_easy_lockscreen);
 
-        FayazSP.init(this);
+        EasylockSP.init(this);
         realPass = getPassword();
         initViews();
 
         status = getIntent().getExtras().getString("passStatus", "check");
         if (status.equals(setStatus))
-            textViewHAHA.setText("Enter a New Password");
+            textViewHAHA.setText(R.string.enter_a_new_password_txt);
+        String disableStatus = "disable";
         if (status.equals(disableStatus)) {
-            FayazSP.put("password", null);
-            Toast.makeText(this, "Password Disabled", Toast.LENGTH_SHORT).show();
+            EasylockSP.put("password", null);
+            Toast.makeText(this, getString(R.string.password_disabled_txt), Toast.LENGTH_SHORT).show();
             gotoActivity();
         }
     }
 
     private void initViews() {
-        textViewHAHA = findViewById(R.id.haha_text);
+        textViewHAHA = findViewById(R.id.span_text);
         textViewDot = findViewById(R.id.dotText);
-        textViewForgotPassword = findViewById(R.id.forgot_pass_textview);
-        buttonTick = findViewById(R.id.lbtnTick);
-        imageButtonDelete = findViewById(R.id.lbtnDelete);
-        relativeLayoutBackground = findViewById(R.id.background_layout);
+        TextView textViewForgotPassword = findViewById(R.id.forgot_pass_textview);
+        Button buttonEnter = findViewById(R.id.lbtnEnter);
+        ImageButton imageButtonDelete = findViewById(R.id.lbtnDelete);
+        RelativeLayout relativeLayoutBackground = findViewById(R.id.background_layout);
         relativeLayoutBackground.setBackgroundColor(EasyLock.backgroundColor);
 
         textViewForgotPassword.setOnClickListener(EasyLock.onClickListener);
 
-        imageButtonDelete.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (passString.length() > 0)
-                    passString = passString.substring(0, passString.length() - 1);
-                textViewDot.setText(passString);
-            }
+        imageButtonDelete.setOnClickListener(view -> {
+            if (passString.length() > 0)
+                passString = passString.substring(0, passString.length() - 1);
+            textViewDot.setText(passString);
         });
 
-        buttonTick.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
+        buttonEnter.setOnClickListener(view -> {
 
-                //
-                if (status.equals(checkStatus)) {
-                    if (passString.equals(realPass)) {
+            //
+            switch (status) {
+                case checkStatus:
+                    if (passString != null && passString.equals(realPass)) {
                         finish();
                     } else {
                         passString = "";
                         textViewDot.setText(passString);
-                        Toast.makeText(LockscreenActivity.this, "Incorrect Password", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(this, getString(R.string.incorrect_password_txt), Toast.LENGTH_SHORT).show();
                     }
-                }
+                    break;
 
                 //
-                else if (status.equals(setStatus)) {
+                case setStatus:
                     //
                     tempPass = passString;
                     passString = "";
                     status = setStatus1;
 
-                    textViewHAHA.setText("Confirm Password");
+                    textViewHAHA.setText(R.string.confirm_password_txt);
                     textViewDot.setText(passString);
-                }
+                    break;
 
                 //
-                else if (status.equals(setStatus1)) {
+                case setStatus1:
                     //
                     if (passString.equals(tempPass)) {
-                        FayazSP.put("password", passString);
-                        Toast.makeText(LockscreenActivity.this, "Password is set", Toast.LENGTH_SHORT).show();
+                        EasylockSP.put("password", passString);
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.password_is_set_txt), Toast.LENGTH_SHORT).show();
                         gotoActivity();
                     } else {
 
@@ -129,43 +124,43 @@ public class LockscreenActivity extends LockscreenHandler implements ActivityCha
                         status = setStatus;
 
                         textViewDot.setText(passString);
-                        textViewHAHA.setText("Enter a New Password");
-                        Toast.makeText(LockscreenActivity.this, "Please Enter a New Password Again", Toast.LENGTH_SHORT).show();
+                        textViewHAHA.setText(R.string.enter_a_new_password_txt);
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.please_enter_a_new_password_again_txt), Toast.LENGTH_SHORT).show();
                     }
-                }
+                    break;
 
                 //
-                else if (status.equals(changeStatus)) {
+                case changeStatus:
                     if (passString.equals(realPass)) {
                         tempPass = passString;
                         passString = "";
                         tempPass = "";
                         status = changeStatus1;
 
-                        textViewHAHA.setText("Enter a New Password");
+                        textViewHAHA.setText(R.string.enter_a_new_password_txt);
                         textViewDot.setText(passString);
                     } else {
                         passString = "";
                         textViewDot.setText(passString);
-                        Toast.makeText(LockscreenActivity.this, "Please Enter Current Password", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.please_enter_current_password_txt), Toast.LENGTH_SHORT).show();
                     }
-                }
+                    break;
 
                 //
-                else if (status.equals(changeStatus1)) {
+                case changeStatus1:
                     tempPass = passString;
                     passString = "";
                     status = changeStatus2;
 
-                    textViewHAHA.setText("Confirm Password");
+                    textViewHAHA.setText(R.string.confirm_password_txt);
                     textViewDot.setText(passString);
-                }
+                    break;
 
                 //
-                else if (status.equals(changeStatus2)) {
+                case changeStatus2:
                     if (passString.equals(tempPass)) {
-                        FayazSP.put("password", passString);
-                        Toast.makeText(LockscreenActivity.this, "Password Changed", Toast.LENGTH_SHORT).show();
+                        EasylockSP.put("password", passString);
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.password_changed_txt), Toast.LENGTH_SHORT).show();
                         gotoActivity();
                     } else {
 
@@ -175,21 +170,21 @@ public class LockscreenActivity extends LockscreenHandler implements ActivityCha
                         status = changeStatus1;
 
                         textViewDot.setText(passString);
-                        textViewHAHA.setText("Enter a New Password");
-                        Toast.makeText(LockscreenActivity.this, "Please Enter a New Password Again", Toast.LENGTH_SHORT).show();
+                        textViewHAHA.setText(R.string.enter_a_new_password_txt);
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.please_enter_a_new_password_again_txt), Toast.LENGTH_SHORT).show();
                     }
-                }
-
+                    break;
             }
+
         });
 
-        for (int i = 0; i < passButtonIds.length; i++) {
-            final Button button = findViewById(passButtonIds[i]);
+        for (int passButtonId : passButtonIds) {
+            final Button button = findViewById(passButtonId);
             button.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
                     if (passString.length() >= 8) {
-                        Toast.makeText(LockscreenActivity.this, "Max 8 characters", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(LockscreenActivity.this, getString(R.string.max_8_characters_txt), Toast.LENGTH_SHORT).show();
                     } else {
                         passString += button.getText().toString();
                     }
@@ -200,7 +195,7 @@ public class LockscreenActivity extends LockscreenHandler implements ActivityCha
     }
 
     private String getPassword() {
-        return FayazSP.getString("password", null);
+        return EasylockSP.getString("password", null);
     }
 
     private void gotoActivity() {
@@ -209,6 +204,7 @@ public class LockscreenActivity extends LockscreenHandler implements ActivityCha
         finish();
     }
 
+    @SuppressWarnings("rawtypes")
     @Override
     public void activityClass(Class activityClassToGo) {
         classToGoAfter = activityClassToGo;
