@@ -3,8 +3,9 @@ package p32929.easypasscodelock.Utils;
 import android.app.ActivityManager;
 import android.content.ComponentCallbacks2;
 import android.content.ComponentName;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.List;
 
@@ -13,25 +14,25 @@ import java.util.List;
  */
 
 public class LockscreenHandler extends AppCompatActivity implements ComponentCallbacks2 {
-    private static boolean wentToBg = false;
+    private static boolean WentToBackground = false;
+    private final String TAG = "EasyLock";
     private String packageName = "";
-    private String TAG = "Fayaz";
 
     @Override
     public void onTrimMemory(int i) {
+        super.onTrimMemory(i);
         ActivityManager am = (ActivityManager) this.getSystemService(ACTIVITY_SERVICE);
         List<ActivityManager.RunningTaskInfo> taskInfo = am.getRunningTasks(1);
-        Log.d("topActivity", "CURRENT Activity ::"
-                + taskInfo.get(0).topActivity.getClassName());
+        Log.d("Activity", "CURRENT Activity ::" + taskInfo.get(0).topActivity.getClassName());
 
-        if (taskInfo != null && taskInfo.size() > 0) {
+        if (taskInfo.size() > 0) {
             ComponentName componentInfo = taskInfo.get(0).topActivity;
             packageName = componentInfo.getPackageName();
         }
         if (!packageName.equals(getPackageName()) && i == ComponentCallbacks2.TRIM_MEMORY_UI_HIDDEN) {
             // We're in the Background
-            wentToBg = true;
-            Log.d(TAG, "wentToBg: " + wentToBg);
+            WentToBackground = true;
+            Log.d(TAG, "WentToBackground: " + WentToBackground);
         }
     }
 
@@ -39,10 +40,10 @@ public class LockscreenHandler extends AppCompatActivity implements ComponentCal
     protected void onResume() {
         super.onResume();
 
-        if (wentToBg && FayazSP.getString("password", null) != null) {
+        if (WentToBackground && EasylockSP.getString("password", null) != null) {
             // We're in the foreground & password != null
-            wentToBg = false;
-            Log.d(TAG, "wentToBg: " + wentToBg);
+            WentToBackground = false;
+            Log.d(TAG, "WentToBackground: " + WentToBackground);
 
             EasyLock.checkPassword(this);
         }
